🚨 🏷️ Fix initial TS errors in `index.ts`
The `input` variable was not considered as an `HTMLInputElement` due to the complex selector used in the `document.querySelector` method call, leading to an error when using the `value` property.
Also, in the `addTodo()` function, the `editMode` property was not explicitly set to `false`, and therefore not matching the inferred todo type.